+++ sonic/database/ent/runtime.go
@@ -291,11 +291,11 @@ func init() {
 	// releaseDescOneTimeFeeAmount is the schema descriptor for OneTimeFeeAmount field.
 	releaseDescOneTimeFeeAmount := releaseFields[6].Descriptor()
 	// release.DefaultOneTimeFeeAmount holds the default value on creation for the OneTimeFeeAmount field.
-	release.DefaultOneTimeFeeAmount = releaseDescOneTimeFeeAmount.Default.(int32)
+	release.DefaultOneTimeFeeAmount = releaseDescOneTimeFeeAmount.Default.(int64)
 	// releaseDescSubscriptionFee is the schema descriptor for SubscriptionFee field.
 	releaseDescSubscriptionFee := releaseFields[7].Descriptor()
 	// release.DefaultSubscriptionFee holds the default value on creation for the SubscriptionFee field.
-	release.DefaultSubscriptionFee = releaseDescSubscriptionFee.Default.(int32)
+	release.DefaultSubscriptionFee = releaseDescSubscriptionFee.Default.(int64)
 	// releaseDescID is the schema descriptor for id field.
 	releaseDescID := releaseFields[0].Descriptor()
 	// release.DefaultID holds the default value on creation for the id field.
